devtools/client/shared/components/test/mochitest/test_frame_01.html

<!DOCTYPE HTML>
<html>
<!--
Test the formatting of the file name, line and columns are correct in frame components,
with optional columns, unknown and non-URL sources.
-->
<head>
  <meta charset="utf-8">
  <title>Frame component test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    let React = browserRequire("devtools/client/shared/vendor/react");
    let Frame = React.createFactory(browserRequire("devtools/client/shared/components/frame"));
    ok(Frame, "Should get Frame");
    let frame;

    // Check when there's a column
    yield checkFrameComponent({
      source: "http://myfile.com/mahscripts.js",
      line: 55,
      column: 10,
    }, {
      file: "mahscripts.js",
      line: 55,
      column: 10,
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js:55:10",
    });

    // Check when there's no column
    yield checkFrameComponent({
      source: "http://myfile.com/mahscripts.js",
      line: 55,
    }, {
      file: "mahscripts.js",
      line: 55,
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js:55",
    });

    // Check when column === 0
    yield checkFrameComponent({
      source: "http://myfile.com/mahscripts.js",
      line: 55,
      column: 0,
    }, {
      file: "mahscripts.js",
      line: 55,
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js:55",
    });

    // Check when there's no parseable URL source;
    // should not render line/columns
    yield checkFrameComponent({
      source: "self-hosted",
      line: 1,
    }, {
      file: "self-hosted",
      shouldLink: false,
      tooltip: "self-hosted",
    });

    // Check when there's no source;
    // should not render line/columns
    yield checkFrameComponent({
      line: 1,
    }, {
      file: "(unknown)",
      shouldLink: false,
      tooltip: "(unknown)",
    });

    // Check when there's a column, but no number;
    // no line/column info should render
    yield checkFrameComponent({
      source: "http://myfile.com/mahscripts.js",
      column: 55,
    }, {
      file: "mahscripts.js",
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js",
    });

    // Check when line is 0; this should be an invalid
    // line option, so don't render line/column
    yield checkFrameComponent({
      source: "http://myfile.com/mahscripts.js",
      line: 0,
      column: 55,
    }, {
      file: "mahscripts.js",
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js",
    });

    // Check when source is via Scratchpad; we should render out the
    // lines and columns as this is linkable.
    yield checkFrameComponent({
      source: "Scratchpad/1",
      line: 10,
      column: 50,
    }, {
      file: "Scratchpad/1",
      line: 10,
      column: 50,
      shouldLink: true,
      tooltip: "View source in Debugger → Scratchpad/1:10:50",
    });

    // Check that line and column can be strings
    yield checkFrameComponent({
      source: "http://myfile.com/mahscripts.js",
      line: "10",
      column: "55",
    }, {
      file: "mahscripts.js",
      line: 10,
      column: 55,
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js:10:55",
    });

    // Check that line and column can be strings,
    // and that the `0` rendering rules apply when they are strings as well
    yield checkFrameComponent({
      source: "http://myfile.com/mahscripts.js",
      line: "0",
      column: "55",
    }, {
      file: "mahscripts.js",
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js",
    });

    function* checkFrameComponent (input, expected) {
      let frame = ReactDOM.render(Frame({
        frame: input,
        onClick: () => {},
      }), window.document.body);
      yield forceRender(frame);
      let source = input.source;
      checkFrameString(Object.assign({ frame, source }, expected));
    }

  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
